Refactor onUpdate Function Inside Job Controller
This commit refactors the onUpdate function within the Job controller,
with a specific focus on reorganzing the various logic initiated by Job
update events in order to establish a "handler" pattern for routing
events to the proper handler functions for processing (as seen in the
native Kubernetes controllers).  This drastically reduces the size of
the onUpdate function (which is currently just a large series of "if"
statements within that single function responsible for detecting and
processing various update events), while establishing a clear pattern
for the routing of various events that can be utilized across the
current controller package.

Changes include:

- The onUpdate function now simply routes the updated job to the proper
handler; it no longer contains any logic itself for processing the
event beyond simply routing it
- Handlers for various job types (pgBackRest, rmdata, pgdump, etc.)
have been separated into separate go files for a cleaner separation of
handler logic based on the various types of Jobs created by the
PostgreSQL Operator
- All go files associated with the job controller (including the job
controller itself), as well as each controller file associated with the
various other controllers (the pod controller, pgcluster controller,
etc.) now been moved into their own packages underneath the
"controller" package.  The idea with this change is to cleanly separate
and logically organize the functionality for each controller into
separate packages. This structure also matches the organizational
structure of the source code for the native controllers developed for
Kubernetes itself.

Please note that while some minor updates were made to the logic
formerly contained within the onUpdate function (either as needed to
support this change, or to simply do some minor cleanup/refactoring),
the primary focus of this update was simply to create a cleaner
organizational structure and pattern for handling controller events.

And finally, certain functionality within the Job onUpdate function
that is no longer utilized by the Operator has been removed.  This
includes logic to process update events for pg_basebackup backups
and restores.
