+++ controller/job/jobutil.go
@@ -1,0 +1,41 @@
+package job
+
+/*
+Copyright 2017 - 2020 Crunchy Data Solutions, Inc.
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import (
+	apiv1 "k8s.io/api/batch/v1"
+	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// isJobSuccessful returns true if the job provided completed successfully.  Otherwise
+// it returns false.  Per the Kubernetes documentation, "the completion time is only set
+// when the job finishes successfully".  Therefore, the presence of a completion time can
+// be utilized to determine whether or not the job was successful.
+func isJobSuccessful(job *apiv1.Job) bool {
+	return job.Status.CompletionTime != nil
+}
+
+// isJobInForegroundDeletion determines if a job is currently being deleted using foreground
+// cascading deletion, as indicated by the presence of value “foregroundDeletion” in the jobs
+// metadata.finalizers.
+func isJobInForegroundDeletion(job *apiv1.Job) bool {
+	for _, finalizer := range job.Finalizers {
+		if finalizer == meta_v1.FinalizerDeleteDependents {
+			return true
+		}
+	}
+	return false
+}
